Created 5-day forecast; made minor CSS updates

+++ src/FormattedWeekday.js
@@ -2,6 +2,6 @@ import React from "react";
 
 export default function FormattedWeekday({ timestamp }) {
   let date = new Date(timestamp * 1000);
-  let options = { weekday: "long" };
+  let options = { weekday: "short" };
   return <div>{date.toLocaleDateString(undefined, options)}</div>;
 }
